app/requirements.txt
Django==3.1.4
djangorestframework==3.12.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.4
requests==2.23.0
sklearn==0.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow-cpu
tensorflow-estimator==2.4.0
pytz==2018.5
gunicorn==19.7.1
whitenoise==5.0.1
h5py==2.10.0
wrapt==1.12.1
grpcio==1.32.0
wheel==0.35
django-heroku